refactor: move ToolOutputTrimmer to extensions/
Move ToolOutputTrimmer up one level from extensions/memory/ to
extensions/ since call_model_input_filter is not tied to the
memory subsystem.

- src/agents/extensions/memory/tool_output_trimmer.py → src/agents/extensions/tool_output_trimmer.py
- tests/extensions/memory/test_tool_output_trimmer.py → tests/extensions/test_tool_output_trimmer.py
- Import path: from agents.extensions import ToolOutputTrimmer
