+++ src/agents/extensions/__init__.py
@@ -1,0 +1,3 @@
+from .tool_output_trimmer import ToolOutputTrimmer
+
+__all__ = ["ToolOutputTrimmer"]
