+++ src/agents/extensions/memory/__init__.py
@@ -22,8 +22,6 @@ if TYPE_CHECKING:
     from .redis_session import RedisSession
     from .sqlalchemy_session import SQLAlchemySession
 
-from .tool_output_trimmer import ToolOutputTrimmer
-
 __all__: list[str] = [
     "AdvancedSQLiteSession",
     "AsyncSQLiteSession",
@@ -33,7 +31,6 @@ __all__: list[str] = [
     "EncryptedSession",
     "RedisSession",
     "SQLAlchemySession",
-    "ToolOutputTrimmer",
 ]
